Dont use small integer fields for ports
We use port numbers exceeding the range for *signed* 16-bit integers (which is what smallint is in postgres).

+++ scionlab/models/vpn.py
@@ -45,7 +45,7 @@ class VPN(models.Model):
         related_name='vpn_servers',
         on_delete=models.CASCADE
     )
-    server_port = models.PositiveSmallIntegerField()
+    server_port = models.PositiveIntegerField()
     subnet = models.CharField(max_length=15)
     server_vpn_ip = models.GenericIPAddressField()
     private_key = models.TextField(null=True, blank=True)
